requirements.txt
fastapi
uvicorn
python-multipart
pytesseract
opencv-python-headless
pymupdf
Pillow
numpy